projects/DWnews/backend/requirements.txt
# The Daily Worker - Backend Dependencies
# Python 3.9+

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9  # PostgreSQL driver
aiosqlite==0.19.0       # Async SQLite driver

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# HTTP & API Clients
httpx==0.26.0
aiohttp==3.9.1
requests==2.31.0

# RSS Feed Parsing
feedparser==6.0.11

# Social Media APIs
tweepy==4.14.0          # Twitter API
praw==7.7.1             # Reddit API

# LLM Clients
anthropic==0.18.0       # Claude API
openai==1.10.0          # OpenAI API
google-generativeai>=1.0.0  # Google Gemini API (Gemini 2.5 Flash Image)

# Payment Processing
stripe==7.10.0          # Stripe SDK for Python

# Email Service
sendgrid==6.11.0        # SendGrid email API (free tier: 100 emails/day)

# Content Processing
beautifulsoup4==4.12.3
lxml==5.1.0
html2text==2024.2.26

# Reading Level Analysis
textstat==0.7.3
nltk==3.8.1

# Image Processing
Pillow==10.2.0
pillow-heif==0.15.0

# Utilities
python-dateutil==2.8.2
pytz==2024.1
validators==0.22.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0  # For testing async clients

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Logging
python-json-logger==2.0.7
